+++ src/dataLayer/updateData.ts
@@ -6,3 +6,8 @@ export const updateObject = async (data: DataObject): Promise<DataObject> => {
   const obj = data;
   return await postData(path, obj).then((data) => data);
 };
+
+export const createNewObject = async (object: any) => {
+  const path = 'create/object';
+  return await postData(path, object);
+};